backend/requirements.txt
fastapi
uvicorn
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.6.1
pydantic-settings==2.1.0
redis==5.0.1
alembic==1.13.1
python-dotenv==1.0.1 
jose
email-validator
google-adk